refactor(ui): tous les composants des dashboards sont en tsx
