--- packages/ui/src/components/_ui/button-plus.vue
@@ -1,9 +1,0 @@
-<template>
-  <button class="btn rnd-xs p-s">
-    <Icon name="plus" size="M" />
-  </button>
-</template>
-
-<script setup lang="ts">
-import { Icon } from '@/components/_ui/icon'
-</script>
